requirements.txt
requests
redis
mysql.connector
plotly
